Refactor: Moved Resume FAB to seperate file
